yeling-AI/VPN/release/source/requirements-pin.txt
flask==2.3.2
requests==2.31.0
PyYAML==6.0
netifaces==0.11.0
